requirements.txt
Flask==3.0.1
Werkzeug==3.0.1
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1
paddlepaddle==3.0.0
paddleocr>=2.8.0
PyMuPDF==1.23.8
Pillow==10.1.0
opencv-python-headless==4.8.1.78
numpy==1.24.3
psutil==5.9.6
